# Performance improvements for the event query API