--- a/libs/pubsub/query/oldquery/fuzz_test/main.go
+++ b/libs/pubsub/query/oldquery/fuzz_test/main.go
@@ -3,7 +3,7 @@ package fuzz_test
 import (
 	"fmt"
 
-	"github.com/tendermint/tendermint/libs/pubsub/query"
+	query "github.com/tendermint/tendermint/libs/pubsub/query/oldquery"
 )
 
 func Fuzz(data []byte) int {