… .github/workflows/coverage.yml
… internal/eventbus/event_bus_test.go
… internal/inspect/inspect_test.go
… internal/state/indexer/block/kv/kv_test.go
… internal/state/indexer/sink/kv/kv_test.go
… internal/state/indexer/tx/kv/kv_bench_test.go
… internal/state/indexer/tx/kv/kv_test.go
… libs/pubsub/example_test.go
… libs/pubsub/pubsub_test.go
… libs/pubsub/query/bench_test.go
… libs/pubsub/query/compile_test.go
… libs/pubsub/query/oldquery/Makefile
… libs/pubsub/query/oldquery/empty.go
… libs/pubsub/query/oldquery/empty_test.go
… libs/pubsub/query/oldquery/parser_test.go
… libs/pubsub/query/oldquery/peg.go
… libs/pubsub/query/oldquery/query.go
… libs/pubsub/query/oldquery/query.peg
… libs/pubsub/query/oldquery/query.peg.go
… libs/pubsub/query/oldquery/query_test.go
… libs/pubsub/query/query_test.go
… libs/pubsub/query/syntax/syntax_test.go
… types/events.go
… types/events_test.go