Rework the implementation of event query parsing and execution to improve performance and reduce memory usage.

Previous memory and CPU profiles of the pubsub service showed query processing as a significant hotspot. While we don't have evidence that this is visibly hurting users, fixing it is fairly easy and self-contained.

Updates #6439.

## Structure

- Move the existing query implementation to the `oldquery` subdirectory. This can probably be deleted before merging.
- Add a `syntax` subpackage providing a lexical scanner and parser to replace the generated PEG parser.
- Pre-compile the query to avoid repeated syntax traversal and allocations during matching.
- Update usage in the rest of the repository.

## Benchmarks

Typical benchmark results comparing the original implementation (PEG) with the reworked implementation (Custom):
```
TEST                        TIME/OP  BYTES/OP  ALLOCS/OP  SPEEDUP   MEM SAVING
BenchmarkParsePEG-12       51716 ns  526832    27
BenchmarkParseCustom-12     2167 ns    4616    17         23.8x     99.1%
BenchmarkMatchPEG-12        3086 ns    1097    22
BenchmarkMatchCustom-12    294.2 ns      64     3         10.5x     94.1%
```