@creachadair 2021-11-25
> closes #6439.
> 
> Amazing work!!

I'm not sure it entirely fixes #6439, quite a lot of memory per subscription is also spent on queuing and buffering messages for the clients. This should help some, though. 🙂